Module Options

A type that represents either a wrapped generic value or no value.

Option types provide a safe way to deal with potentially absent values.

Option Type

type nonrec 'a option = 'a option =

| None
No value.

| Some of 'a
Some generic value.

A type for generic optional values.

This type represents an optional value: it is either Some x and contains the value x, or None, and does not contain any value. The complementary Option module includes operations on options.

Note: This is an alias for the builtin option type included here for documentation.

module Option : sig ... end

Module with operations for the option type.

Public Definitions

These are the public definitions that will be exported when the top-level Options module is open.

val some : 'a -> 'a Option.t

Produces an option containing a given value.

some x wraps x with the Some constructor. This function is useful for partial application.

Examples

assert (List.map some [1; 2; 3] = [Some 1; Some 2; Some 3])

val none : 'a Option.t

none is the None constructor. Not very useful but included for consistency with some.

val is_some : 'a Option.t -> bool

is_some option is true if option is a Some value.

val is_none : 'a Option.t -> bool

is_none option is true if option is a None value.

val if_some : ('a -> unit) -> 'a Option.t -> unit

if_some f option applies an effectful function f to the value wrapped in option or does nothing if option contains no value.

Examples

Some "Hello, world!" |> if_some print_endline
(* Output: Hello, World! *)

None |> if_some print
(* No output *)

val if_none : (unit -> unit) -> 'a Option.t -> unit

if_none f option calls an effectful function f if option is a None value or does nothing otherwise.

Examples

None |> if_none (fun () -> print_endline "no value")
(* Output: no value *)

Some 42 |> if_none (fun () -> print_endline "no value")
(* No output *)

val or_else : (unit -> 'a) -> 'a Option.t -> 'a

or_else f opt extracts the optional value. If the optional is None, the default value f () is returned. A thunk is used instead of a direct value to avoid the default value evaluation when the option is set.

See also: Option.(or)

Examples

assert (Some 1 |> Option.or_else (fun () -> 0) = 1);
assert (None |> Option.or_else (fun () -> 0) = 0);

(* [read_line] will not be called in the following example. *)
assert (Some "Bob" |> Option.or_else read_line = "Bob");

val or : 'a Option.t -> 'a -> 'a

Unwrap the option or use a default value.

option or default is the flipped infix version of Option.or_else equivalent to option |> or_else (fun () -> default). The default value will be always evaluated, for lazy evaluation use Option.or_else.

Note: this operator has a very low precedence, e.g. option or 0 + 1 is evaluated as option or (0 + 1) and not as (option or 0) + 1. Complex expressions should be grouped explicitly to avoid ambiguity.

See also: Option.or_else

Examples

assert ((None or 0) = 0);
assert ((None or 1 + 2) = 3);
assert ((Some 1 or 0) = 1);
assert ((Some 1 or 1 + 2) = 4);
assert (((Some 1 or 1) + 2) = 3);