Module Options.Option

Module with operations for the option type.

Option types are very useful to represent computations that may fail in one obvious way (like getting an element from an empty list). For more complicated errors consider using the result type which can include detailed descriptions of the errors.

type nonrec 'a t = 'a option =

| None
No value.

| Some of 'a
Some generic value.

Optional values of type 'a.

module Prelude : sig ... end

Module exports suitable for opening.

Public Exports

val some : 'a -> 'a t

Produces an option containing a given value.

some x wraps x with the Some constructor. This function is useful for partial application.

Examples

assert (List.map some [1; 2; 3] = [Some 1; Some 2; Some 3])

val none : 'a t

none is the None constructor. Not very useful but included for consistency with some.

val is_some : 'a t -> bool

is_some option is true if option is a Some value.

val is_none : 'a t -> bool

is_none option is true if option is a None value.

val if_some : ('a -> unit) -> 'a t -> unit

if_some f option applies an effectful function f to the value wrapped in option or does nothing if option contains no value.

Examples

Some "Hello, world!" |> if_some print_endline
(* Output: Hello, World! *)

None |> if_some print
(* No output *)

val if_none : (unit -> unit) -> 'a t -> unit

if_none f option calls an effectful function f if option is a None value or does nothing otherwise.

Examples

None |> if_none (fun () -> print_endline "no value")
(* Output: no value *)

Some 42 |> if_none (fun () -> print_endline "no value")
(* No output *)

val or_else : (unit -> 'a) -> 'a t -> 'a

or_else f opt extracts the optional value. If the optional is None, the default value f () is returned. A thunk is used instead of a direct value to avoid the default value evaluation when the option is set.

See also: Option.(or)

Examples

assert (Some 1 |> Option.or_else (fun () -> 0) = 1);
assert (None |> Option.or_else (fun () -> 0) = 0);

(* [read_line] will not be called in the following example. *)
assert (Some "Bob" |> Option.or_else read_line = "Bob");

val or : 'a t -> 'a -> 'a

Unwrap the option or use a default value.

option or default is the flipped infix version of Option.or_else equivalent to option |> or_else (fun () -> default). The default value will be always evaluated, for lazy evaluation use Option.or_else.

Note: this operator has a very low precedence, e.g. option or 0 + 1 is evaluated as option or (0 + 1) and not as (option or 0) + 1. Complex expressions should be grouped explicitly to avoid ambiguity.

See also: Option.or_else

Examples

assert ((None or 0) = 0);
assert ((None or 1 + 2) = 3);
assert ((Some 1 or 0) = 1);
assert ((Some 1 or 1 + 2) = 4);
assert (((Some 1 or 1) + 2) = 3);

val case : some:('a -> 'b) -> none:(unit -> 'b) -> 'a t -> 'b

Pattern-matching on option values with functions.

case ~none ~some option is the function some applied to the value wrapped by option, or none () if option has no value.

Examples

assert (None    |> Option.case ~some:((+) 1) ~none:(fun () -> 0) = 0);
assert (Some 42 |> Option.case ~some:((+) 1) ~none:(fun () -> 0) = 43);

val is_empty : 'a t -> bool

is_empty opt is true if the option is None and false otherwise.

val catch : (unit -> 'a) -> 'a t

catch f wraps the call to f, returning None if it raises an exception or its return value as Some otherwise.

Examples

(* read_line may raise End_of_file. *)
Option.catch read_line |> Option.or_else (fun () -> "nothing")

val hash : ('a -> int) -> 'a t -> int

Produces a hash for the option value.

Conversions

val to_bool : 'a t -> bool

to_bool self is an alis for is_some.

Examples

assert (Option.to_bool (Some 42) = true);
assert (Option.to_bool None = false);

val to_list : 'a t -> 'a list

to_list self is a singleton list with the value wrapped by self or an empty list if self is None.

Examples

assert (Option.to_bool (Some 42) = true);
assert (Option.to_bool None = false);

val to_result : error:'b -> 'a t -> ('a, 'b) Stdlib.result

to_result ~error t converts an option value into a result. If the option value is None, the error value will be used for the Error result case.

Examples

assert (Option.to_result ~error:(Some 42) = Ok 42);
assert (Option.to_result ~error:"No" None = Error "No");

Comparison and equality

val compare : ('a -> 'a -> [ `Less | `Equal | `Greater ]) -> 'a t -> 'a t -> [ `Less | `Equal | `Greater ]

Comparison for option values. See Compat.compare for the integer-based implementation.

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

Checks if two option values are equal.

Monad instance

val return : 'a -> 'a t

val bind : ('a -> 'b t) -> 'a t -> 'b t

Functor instance

val map : ('a -> 'b) -> 'a t -> 'b t

val (<@>) : ('a -> 'b) -> 'a t -> 'b t

map f self will apply f to the value wrapped by self, returning an option with the resulting value, or None if self does not not have any value. This operator is an infix version of map.

Note: <@> (read as "apply") is the infix version of map.

Examples

assert (Option.map List.reverse (Some [1; 2; 3]) = Some [3; 2; 1]);
assert (Int.to_string <@> None = None);

Unsafe Operations

This section includes the operations that may raise exceptions.

exception No_value

Exception raised when forcing a None option value.

val force : 'a t -> 'a

Forces the extraction the optional value.

Examples

assert (Option.force (List.head [1; 2; 3]) = 1);
assert (raises (Option.force (List.head [])))

raises No_value

if option is None.

val or_fail : string -> 'a t -> 'a

or_fail message option forces the extraction of the optional value and fails with message if option does not contain a value.

Examples

assert (List.head [1; 2; 3] |> Option.or_fail "empty list" = 1);

raises Failure

if self is None.

module Compat : sig ... end